Fix requested changes.
